+++ plivo/resources/__init__.py
@@ -21,3 +21,4 @@ from .campaign import Campaign
 from .profile import Profile
 from .multipartycall import MultiPartyCalls, MultiPartyCall, MultiPartyCallParticipant
 from .verify import Sessions
+from .tollfree_verification import TollfreeVerifications